Moved photograph into the name section as per the designs.
