+++ packages/esm-patient-registration-app/src/patient-registration/patient-registration.component.tsx
@@ -267,9 +267,11 @@ export const PatientRegistration: React.FC<PatientRegistrationProps> = ({ savePa
                         values: props.values,
                         inEditMode,
                         setFieldValue: props.setFieldValue,
+                        setCapturePhotoProps,
+                        currentPhoto,
                       }}>
                       {sections.map((section, index) => (
-                        <div key={index}>{getSection(section, index, setCapturePhotoProps, currentPhoto)}</div>
+                        <div key={index}>{getSection(section, index)}</div>
                       ))}
                     </PatientRegistrationContext.Provider>
                   </Grid>
